+++ src/Error.php
@@ -11,7 +11,7 @@
      * @namespace PHPSandbox
      *
      * @author  Elijah Horton <fieryprophet@yahoo.com>
-     * @version 1.3.3
+     * @version 1.3.4
      */
     class Error extends \Exception {
         /* START ERROR CODES */